refactor: extract create folder dialog to a separate component
